Refactor discovery: module tree -> _module_tree; config utils -> _utils
