+++ src_py/docnote_extract/_summarization.py
@@ -31,9 +31,9 @@ from docnote_extract._crossrefs import has_crossreffed_base
 from docnote_extract._crossrefs import has_crossreffed_metaclass
 from docnote_extract._crossrefs import is_crossreffed
 from docnote_extract._extraction import ModulePostExtraction
+from docnote_extract._module_tree import ModuleTreeNode
 from docnote_extract._types import Singleton
-from docnote_extract.discovery import ModuleTreeNode
-from docnote_extract.discovery import validate_config
+from docnote_extract._utils import validate_config
 from docnote_extract.filtering import ModuleObjectFilter
 from docnote_extract.filtering import ObjectFilter
 from docnote_extract.filtering import filter_inclusion_rules
